+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 更新日志
 
+### 1.2.30
+- ✅ 将 `main.py` 中不依赖插件实例状态的工具方法迁移到 `src/tool.py`，保持入口类聚焦于插件编排
+
 ### 1.2.29
 - ✅ 排除唤醒/命令消息的任意消息精确判题，避免 `/help`、`/ccl` 等命令文本包含答案时被误判为答对
