+++ metadata.yaml
@@ -1,7 +1,7 @@
 name: astrbot_plugin_mrfzccl
 display_name: "明日方舟猜猜乐"
 desc: "明日方舟干员立绘猜猜乐，支持排行榜/名片/比赛模式"
-version: v1.2.29
+version: v1.2.30
 author: lishining
 repo: https://github.com/Li-shi-ling/astrbot_plugin_mrfzccl
 support_platforms:
