+++ src/tool.py
@@ -1,9 +1,14 @@
 import os
 import re
 import json
+import asyncio
+import random
 from pathlib import Path
+from io import BytesIO
 from datetime import datetime
 from collections import Counter
+from PIL import Image
+import numpy as np
 from pypinyin import lazy_pinyin, Style
 from typing import Any, Callable, Iterable, Mapping, Optional
 
@@ -475,6 +480,165 @@ def has_active_game(player: Mapping[str, Any], user_id: str) -> bool:
     """检查用户是否有活跃游戏"""
     data = (player or {}).get(user_id)
     return bool(data and data.get("status") == "active")
+
+
+# 安全取消异步任务。
+def safe_cancel_task(task: asyncio.Task | None) -> None:
+    if not task:
+        return
+    try:
+        if not task.done():
+            task.cancel()
+    except Exception:
+        pass
+
+
+# 解析 LLM 判题输出为布尔结果。
+def parse_llm_judge_result(completion_text: str) -> bool | None:
+    text = str(completion_text or "").strip().lower()
+    if not text:
+        return None
+    text = re.sub(r"^[`\"'\s]+|[`\"'\s]+$", "", text)
+    text = re.sub(r"[.。！？?!]+$", "", text)
+    if text == "true":
+        return True
+    if text == "false":
+        return False
+    if re.fullmatch(r'\{\s*"(result|answer|correct)"\s*:\s*true\s*\}', text):
+        return True
+    if re.fullmatch(r'\{\s*"(result|answer|correct)"\s*:\s*false\s*\}', text):
+        return False
+    return None
+
+
+# 转换为绝对路径。
+def get_absolute_path(path: str) -> str:
+    if not path:
+        raise ValueError("路径不能为空")
+    return os.path.abspath(path)
+
+
+# 从字节内容加载并校验图片。
+def load_image_from_bytes(content: bytes) -> Image.Image:
+    image = Image.open(BytesIO(content))
+    if image.format not in ["JPEG", "PNG", "GIF", "WEBP", "BMP"]:
+        raise Exception(f"不支持的图片格式: {image.format}")
+    image.load()
+
+    width, height = image.size
+    if width > 5000 or height > 5000:
+        raise Exception(f"图片尺寸过大: {width}x{height}")
+    return image
+
+
+# 提取并清理用户输入。
+def extract_and_sanitize_input(text: str, keyword: str) -> str:
+    if not text or not keyword:
+        return ""
+    pattern = rf"{re.escape(keyword)}\s*(.*)"
+    match = re.search(pattern, text)
+    if not match:
+        return ""
+    user_input = match.group(1).strip()
+    cleaned = re.sub(r"[^\u4e00-\u9fa5a-zA-Z0-9\s]", "", user_input)
+    if len(cleaned) > 50:
+        cleaned = cleaned[:50]
+    return cleaned
+
+
+# 生成随机方块遮罩图片。
+def mask_image_with_random_blocks(
+    image: Image.Image,
+    block_count: int = 5,
+    mask_color: tuple[int, int, int] = (0, 0, 0),
+    min_width_percent: int = 10,
+    max_width_percent: int = 20,
+    min_height_percent: int = 10,
+    max_height_percent: int = 20,
+    min_gap_percent: int = 2,
+    avoid_edges: bool = True,
+) -> tuple[Image.Image, list[tuple[int, int, int, int]]]:
+    if image.mode != "RGBA":
+        original_rgba = image.convert("RGBA")
+    else:
+        original_rgba = image.copy()
+
+    width, height = original_rgba.size
+    arr = np.array(original_rgba)
+
+    mask_layer = np.zeros_like(arr)
+    mask_layer[..., 0] = mask_color[0]
+    mask_layer[..., 1] = mask_color[1]
+    mask_layer[..., 2] = mask_color[2]
+    mask_layer[..., 3] = 255
+
+    min_width = max(5, int(width * min_width_percent / 100))
+    max_width = max(min_width, int(width * max_width_percent / 100))
+    min_height = max(5, int(height * min_height_percent / 100))
+    max_height = max(min_height, int(height * max_height_percent / 100))
+    min_gap = int(min(width, height) * min_gap_percent / 100)
+    edge_margin = min_gap if avoid_edges else 0
+
+    blocks: list[tuple[int, int, int, int]] = []
+
+    for _ in range(block_count):
+        for _attempt in range(100):
+            w = random.randint(min_width, max_width)
+            h = random.randint(min_height, max_height)
+            max_x = width - w - edge_margin
+            max_y = height - h - edge_margin
+            if max_x <= edge_margin or max_y <= edge_margin:
+                break
+
+            x1 = random.randint(edge_margin, max_x)
+            y1 = random.randint(edge_margin, max_y)
+            x2, y2 = x1 + w, y1 + h
+
+            conflict = False
+            for bx1, by1, bx2, by2 in blocks:
+                if not (
+                    x2 + min_gap < bx1
+                    or x1 > bx2 + min_gap
+                    or y2 + min_gap < by1
+                    or y1 > by2 + min_gap
+                ):
+                    conflict = True
+                    break
+
+            if not conflict:
+                blocks.append((x1, y1, x2, y2))
+                mask_layer[y1:y2, x1:x2, 3] = 0
+                break
+
+    alpha = mask_layer[..., 3:4] / 255.0
+    result_arr = arr * (1 - alpha) + mask_layer * alpha
+    result_arr = result_arr.astype(np.uint8)
+    result = Image.fromarray(result_arr, "RGBA")
+    return result, blocks
+
+
+# 按比例缩放图片并保持宽高比。
+def resize_to_target(image: Image.Image, target_size: int) -> Image.Image:
+    if target_size <= 0:
+        target_size = 800
+    w, h = image.size
+    if w >= h:
+        new_w = target_size
+        new_h = int(target_size * h / w)
+    else:
+        new_h = target_size
+        new_w = int(target_size * w / h)
+    new_w = max(new_w, 100)
+    new_h = max(new_h, 100)
+    return image.resize((new_w, new_h), Image.Resampling.LANCZOS)
+
+
+# 将 PIL 图片转换为字节。
+def pil_image_to_bytes(image: Image.Image, format: str = "PNG") -> bytes:
+    buf = BytesIO()
+    image.save(buf, format=format, optimize=True)
+    return buf.getvalue()
+
 
 # 清洗ffc消息,转变为指令
 def normalize_compact_fc_command(message_str: str) -> str | None:
